Code example needs string parameter
The code example given in this useContractWrite documentation file didn't work for me unless I added the function name as a string parameter. I'm not sure why it's not included in the documentation, maybe it's been changed in a recent update or maybe it's assumed the user knows to do this but it was not immediately evident to me.