Fix an issue with useContractInfiniteReads with multicall.
The number of results per page was not taking into account the number of calls. As a result, if the number of calls made exceeds the page size, it failed the getNextPageParam check, which made wagmi think there were no more pages.

This fix calculates the number of expected calls per page, and uses that in getNextPageParam.

There was an existing test for this, but the inline snapshot for it was incorrect - this fixes this.