feat: Soften duplicate contract name in the case contracts have identical ABIs
This is useful for instance when contracts appear as duplicates when
ingesting from a single Foundry project that depends on others foundry
projects.

In this case `getArtifactPaths()` in the Foundry plugin pick up the same
contracts twice since they end up existing in both `out/` directories. I
don't think there is a reliable to heuristic to de-duplicate them at
this level.

It seems like duck-typing contracts by the ABI should be fine, it could
always be enabled via an option.

Signed-off-by: Silas Davis <silas@inco.network>