Update @walletconnect/ethereum-provider to version 2.22.4
This version includes a fix (that came in 2.22.2) that addresses an issue related to esm/cjs importing of the logger used by the @walletconnect/ethereum-provider package. This issue resulted in the walletConnect.ts:276 `EthereumProvider.init` call to fail. This subsequently resulted in dependent packages (ie. family/connectkit) to fail when trying to initialize the walletConnect connector.